[JSDK-2] Changes after the code review - remove @Deprecated marker and move the enum Protocol to the package Model
