+++ src/com/mambu/apisdk/model/Protocol.java
@@ -1,4 +1,4 @@
-package com.mambu.apisdk;
+package com.mambu.apisdk.model;
 
 /**
  * Enum that stores application protocols that can be used within the project.
